Rename services namespace to service
