+++ app/api/auth/controller.go
@@ -6,7 +6,7 @@ import (
 	"github.com/gghcode/go-gin-starterkit/app/api/common"
 	"github.com/gghcode/go-gin-starterkit/app/api/user"
 	"github.com/gghcode/go-gin-starterkit/config"
-	"github.com/gghcode/go-gin-starterkit/services"
+	"github.com/gghcode/go-gin-starterkit/service"
 	"github.com/gin-gonic/gin"
 )
 
@@ -17,7 +17,7 @@ type Controller struct {
 }
 
 // NewController return new auth controller instance.
-func NewController(conf config.Configuration, userRepo user.Repository, passport services.Passport) *Controller {
+func NewController(conf config.Configuration, userRepo user.Repository, passport service.Passport) *Controller {
 	return &Controller{
 		conf:    conf.Jwt,
 		service: NewService(conf, userRepo, passport),
