+++ app/api/auth/service.go
@@ -7,23 +7,23 @@ import (
 	"github.com/dgrijalva/jwt-go"
 	"github.com/gghcode/go-gin-starterkit/app/api/user"
 	"github.com/gghcode/go-gin-starterkit/config"
-	"github.com/gghcode/go-gin-starterkit/services"
+	"github.com/gghcode/go-gin-starterkit/service"
 )
 
-// Service is auth service.
+// Service is auth authService.
 type Service interface {
 	VerifyAuthentication(username, password string) (user.User, error)
 	GenerateAccessToken(userID int64) (string, error)
 	IssueRefreshToken(userID int64) (string, error)
 }
 
-// NewService return new auth service instance.
+// NewService return new auth authService instance.
 func NewService(
 	conf config.Configuration,
 	userRepo user.Repository,
-	passport services.Passport) Service {
+	passport service.Passport) Service {
 
-	return &service{
+	return &authService{
 		secretKeyBytes:      []byte(conf.Jwt.SecretKey),
 		accessExpiresInSec:  time.Duration(conf.Jwt.AccessExpiresInSec),
 		refreshExpiresInSec: time.Duration(conf.Jwt.RefreshExpiresInSec),
@@ -32,37 +32,37 @@ func NewService(
 	}
 }
 
-type service struct {
+type authService struct {
 	secretKeyBytes      []byte
 	accessExpiresInSec  time.Duration
 	refreshExpiresInSec time.Duration
 
 	userRepo user.Repository
-	passport services.Passport
+	passport service.Passport
 }
 
-func (service *service) VerifyAuthentication(username, password string) (user.User, error) {
-	loginUser, err := service.userRepo.GetUserByUserName(username)
+func (authService *authService) VerifyAuthentication(username, password string) (user.User, error) {
+	loginUser, err := authService.userRepo.GetUserByUserName(username)
 	if err != nil {
 		return user.EmptyUser, err
 	}
 
-	if !service.passport.IsValidPassword(password, loginUser.PasswordHash) {
+	if !authService.passport.IsValidPassword(password, loginUser.PasswordHash) {
 		return user.EmptyUser, ErrInvalidPassword
 	}
 
 	return loginUser, nil
 }
 
-func (service *service) GenerateAccessToken(userID int64) (string, error) {
+func (authService *authService) GenerateAccessToken(userID int64) (string, error) {
 	claims := &jwt.StandardClaims{
-		ExpiresAt: time.Now().Add(service.accessExpiresInSec * time.Second).Unix(),
+		ExpiresAt: time.Now().Add(authService.accessExpiresInSec * time.Second).Unix(),
 		IssuedAt:  time.Now().Unix(),
 		Subject:   strconv.FormatInt(userID, 10),
 	}
 
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-	tokenString, err := token.SignedString(service.secretKeyBytes)
+	tokenString, err := token.SignedString(authService.secretKeyBytes)
 	if err != nil {
 		return "", err
 	}
@@ -70,15 +70,15 @@ func (service *service) GenerateAccessToken(userID int64) (string, error) {
 	return tokenString, nil
 }
 
-func (service *service) IssueRefreshToken(userID int64) (string, error) {
+func (authService *authService) IssueRefreshToken(userID int64) (string, error) {
 	claims := &jwt.StandardClaims{
-		ExpiresAt: time.Now().Add(service.refreshExpiresInSec * time.Second).Unix(),
+		ExpiresAt: time.Now().Add(authService.refreshExpiresInSec * time.Second).Unix(),
 		IssuedAt:  time.Now().Unix(),
 		Subject:   strconv.FormatInt(userID, 10),
 	}
 
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-	tokenString, err := token.SignedString(service.secretKeyBytes)
+	tokenString, err := token.SignedString(authService.secretKeyBytes)
 	if err != nil {
 		return "", err
 	}
